pyproject.toml
[tool.poetry]
name = "restninja"
version = "0.1.0"
description = ""
authors = ["Hazuli Fidastian <hazulifidastian@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django-ninja = "^0.17.0"
orjson = "^3.6.6"
PyYAML = "^6.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
